requirements.txt
anyio==3.7.1
asn1crypto==0.24.0
certifi==2019.6.16
cffi==1.12.3
cryptography==2.7
dataclass-wizard==0.22.2
exceptiongroup==1.1.2
future==0.17.1
fuzzywuzzy==0.17.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
pycparser==2.19
python-dotenv==0.21.1
python-Levenshtein==0.12.0
python-telegram-bot==20.3
result==0.11.0
six==1.12.0
sniffio==1.3.0
tornado==6.0.3
typing_extensions==4.7.1
